Global player

+++ src/components/SettingsChangers/DeleteAccount.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useRef, useState } from "react";
 import Button from "../UI/Button";
 import styles from "./DeleteAccount.module.css";
 import { motion } from "framer-motion";
@@ -10,24 +10,48 @@ import { playlistActions } from "../../redux/playlist-slice";
 import { deleteCurrentUser } from "../../redux/Actions/loginActions";
 import { updateActions } from "../../redux/update-slice";
 import { useHistory } from "react-router-dom";
+import { resetPlayer } from "../../redux/Actions/playerActions";
+import Warning from "../UI/Warning";
+let modalManageState = false;
+
 export default function DeleteAccount(props) {
   const history = useHistory();
   const dispatch = useDispatch();
+  const passwordInputRef = useRef(null);
+  const audio = useSelector((state) => state.player.audio);
   const [showModal, setShowModal] = useState(false);
   const currentUser = useSelector((state) => state.authentication.currentUser);
+  const [warning, setWarning] = useState(false);
   function toggleModalHandler() {
-    setShowModal((prev) => !prev);
+    if (!modalManageState) {
+      setShowModal((prev) => !prev);
+      setWarning(false);
+    }
+
+    if (modalManageState) {
+      modalManageState = false;
+    }
   }
   function deleteAccount() {
-    dispatch(authActions.deleteAccount(currentUser));
-    sessionStorage.setItem("isLogged", "false");
-    sessionStorage.removeItem("currentUser");
-    dispatch(songsActions.resetSongList());
-    dispatch(playlistActions.resetPlaylists());
-    deleteCurrentUser(currentUser);
-    dispatch(updateActions.shouldUpdate());
-    history.push("/Musify");
+    if (passwordInputRef.current.value === currentUser.password) {
+      dispatch(authActions.deleteAccount(currentUser));
+      sessionStorage.setItem("isLogged", "false");
+      sessionStorage.removeItem("currentUser");
+      dispatch(songsActions.resetSongList());
+      dispatch(playlistActions.resetPlaylists());
+      deleteCurrentUser(currentUser);
+      dispatch(updateActions.shouldUpdate());
+      history.push("/Musify");
+      dispatch(resetPlayer(audio));
+      modalManageState = false;
+      setWarning(false);
+      toggleModalHandler();
+    } else {
+      setWarning(true);
+      modalManageState = true;
+    }
   }
+
   return (
     <motion.div
       key={props.animate}
@@ -43,7 +67,13 @@ export default function DeleteAccount(props) {
           removeItem={deleteAccount}
           closeModal={toggleModalHandler}
           text="Are you sure You want to delete your account?"
-        />
+        >
+          <div className={styles["password-confirmation"]}>
+            <label htmlFor="confirm">Enter your password.</label>
+            <input ref={passwordInputRef} id="confirm"></input>
+            {warning && <Warning>Wrong password.</Warning>}
+          </div>
+        </AreYouSureModal>
       )}
     </motion.div>
   );
